+++ test/controllers/research_topics_controller_test.rb
@@ -83,17 +83,17 @@ class ResearchTopicsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should redirect to intro action for no_votes user" do
-    login(@no_votes_user)
-    get :index
-    assert_redirected_to intro_research_topics_path
-  end
+  # test "should redirect to intro action for no_votes user" do
+  #   login(@no_votes_user)
+  #   get :index
+  #   assert_redirected_to intro_research_topics_path
+  # end
 
-  test "should redirect to first_topics action for novice_user" do
-    login(@novice_user)
-    get :index
-    assert_redirected_to first_topics_research_topics_path
-  end
+  # test "should redirect to first_topics action for novice_user" do
+  #   login(@novice_user)
+  #   get :index
+  #   assert_redirected_to first_topics_research_topics_path
+  # end
 
   # Intro : page explaining rank the research and allowing user to press "get started"
   test "should get intro and redirect for logged out user" do
@@ -113,23 +113,23 @@ class ResearchTopicsControllerTest < ActionController::TestCase
     assert_redirected_to research_topics_path
   end
 
-  test "should get first topics for novice user" do
-    login(@novice_user)
-    get :first_topics
-    assert_response :success
-  end
+  # test "should get first topics for novice user" do
+  #   login(@novice_user)
+  #   get :first_topics
+  #   assert_response :success
+  # end
 
-  test "should get first topics for no_votes user that read the intro" do
-    login(@no_votes_user)
-    get :first_topics, read_intro: 1
-    assert_response :success
-  end
+  # test "should get first topics for no_votes user that read the intro" do
+  #   login(@no_votes_user)
+  #   get :first_topics, read_intro: 1
+  #   assert_response :success
+  # end
 
-  test "should redirect no_votes user to intro if they haven't read the intro" do
-    login(@no_votes_user)
-    get :first_topics
-    assert_redirected_to intro_research_topics_path
-  end
+  # test "should redirect no_votes user to intro if they haven't read the intro" do
+  #   login(@no_votes_user)
+  #   get :first_topics
+  #   assert_redirected_to intro_research_topics_path
+  # end
 
   test "should redirect from first topics to index for experienced users" do
     login(@experienced_user)
@@ -207,20 +207,20 @@ class ResearchTopicsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should not create research topic as novice user" do
-    login(@novice_user)
+  # test "should not create research topic as novice user" do
+  #   login(@novice_user)
 
-    assert_no_difference('ResearchTopic.count') do
-      post :create, research_topic: { text: 'Some new research topic', description: 'Why I think this is important'}
-    end
-  end
+  #   assert_no_difference('ResearchTopic.count') do
+  #     post :create, research_topic: { text: 'Some new research topic', description: 'Why I think this is important'}
+  #   end
+  # end
 
-  test "should not create research topic as no votes user" do
-    login(@no_votes_user)
-    assert_no_difference('ResearchTopic.count') do
-      post :create, research_topic: { text: 'Some new research topic', description: 'Why I think this is important'}
-    end
-  end
+  # test "should not create research topic as no votes user" do
+  #   login(@no_votes_user)
+  #   assert_no_difference('ResearchTopic.count') do
+  #     post :create, research_topic: { text: 'Some new research topic', description: 'Why I think this is important'}
+  #   end
+  # end
 
   test "should not create research topic as logged out user" do
     assert_no_difference('ResearchTopic.count') do
@@ -239,17 +239,17 @@ class ResearchTopicsControllerTest < ActionController::TestCase
     assert_response :success
   end
 
-  test "should get my topics and redirect to intro action for no_votes user" do
-    login(@no_votes_user)
-    get :my_research_topics
-    assert_redirected_to intro_research_topics_path
-  end
+  # test "should get my topics and redirect to intro action for no_votes user" do
+  #   login(@no_votes_user)
+  #   get :my_research_topics
+  #   assert_redirected_to intro_research_topics_path
+  # end
 
-  test "should get my topics and redirect to first_topics action for novice_user" do
-    login(@novice_user)
-    get :my_research_topics
-    assert_redirected_to first_topics_research_topics_path
-  end
+  # test "should get my topics and redirect to first_topics action for novice_user" do
+  #   login(@novice_user)
+  #   get :my_research_topics
+  #   assert_redirected_to first_topics_research_topics_path
+  # end
 
   test "should get my topics and redirect to index for logged out user" do
     get :my_research_topics
@@ -281,30 +281,30 @@ class ResearchTopicsControllerTest < ActionController::TestCase
     assert_equal comment, research_topics(:rt2).topic.posts.last.description
   end
 
-  test "should vote for only the seeded research topics as a no_votes user" do
-    login(@no_votes_user)
-    rt = ResearchTopic.where(category: "seeded").first
+  # test "should vote for only the seeded research topics as a no_votes user" do
+  #   login(@no_votes_user)
+  #   rt = ResearchTopic.where(category: "seeded").first
 
-    assert_no_difference "Vote.count" do
-      post :vote, research_topic_id: @rt2.id, "endorse_#{research_topics(:rt2).id}" => 1
-    end
-    assert_difference "Vote.count" do
-      post :vote, research_topic_id: rt.id, "endorse_#{rt.id}" => 1
-    end
-  end
+  #   assert_no_difference "Vote.count" do
+  #     post :vote, research_topic_id: @rt2.id, "endorse_#{research_topics(:rt2).id}" => 1
+  #   end
+  #   assert_difference "Vote.count" do
+  #     post :vote, research_topic_id: rt.id, "endorse_#{rt.id}" => 1
+  #   end
+  # end
 
 
-  test "should vote for only the seeded research topics as a novice user" do
-    login(@novice_user)
-    rt = ResearchTopic.where(category: "seeded").first
+  # test "should vote for only the seeded research topics as a novice user" do
+  #   login(@novice_user)
+  #   rt = ResearchTopic.where(category: "seeded").first
 
-    assert_no_difference "Vote.count" do
-      post :vote, research_topic_id: @rt2.id, "endorse_#{research_topics(:rt2).id}" => 1
-    end
-    assert_difference "Vote.count" do
-      post :vote, research_topic_id: rt.id, "endorse_#{rt.id}" => 1
-    end
-  end
+  #   assert_no_difference "Vote.count" do
+  #     post :vote, research_topic_id: @rt2.id, "endorse_#{research_topics(:rt2).id}" => 1
+  #   end
+  #   assert_difference "Vote.count" do
+  #     post :vote, research_topic_id: rt.id, "endorse_#{rt.id}" => 1
+  #   end
+  # end
 
   test "should add remote votes for research topics" do
     login(@novice_user)