+++ CHANGELOG.md
@@ -14,6 +14,8 @@
     post if one exists
 - **Landing Page Changes**
   - Blog posts are now viewable on the landing page
+- **Research Topics**
+  - Removed requirement to do a research tutorial before casting votes
 - **Gem Changes**
   - Updated to Ruby 2.3.1
   - Updated to devise 4.1.0